refactor: split qol.rs into themed submodules
Break the 1220-line qol.rs into a qol/ module directory, one cohesive
feature area per file (starting_state, overworld_map, big_q, cards,
hammer_breaks, canoe, beta, macobra). Pure code movement: ROM output is
byte-for-byte identical across seeds and the full flag matrix, all tests
pass, clippy clean.
